Got basic gmock test compiling
Created a failing tests case for testing ImagePng constructor, which
pulls in mock classes for libpng and file calls. Decided to *not*
include gmock.h in the librend/external_interfaces mock files, as I do
not want to add gmock as a dependency of the library. gmock should only
be a dependency of the test app.