Wireframe rendering complete!
The main program now accepts a wavefront obj file name as the first
command line argument, and will render a wireframe image of the object
to ./foo.png. The model is scaled appropriately to fit entirely within
the 1000x1000 png.

Added a CMake download job to automatically clone ssloy/tinyrenderer, as
that repo contains nice sample objects to use with this tutorial.

Moved width and height attributes from PngImage to the ImageInterface
(base) class. All images have a width and height, right?

Reordered the horizontal rows when writing to png images such that
(x,y)=(0,0) is at the bottom left.

Added Vertex and Face classes, and guts of Model class.

Faces are loaded from a file and are assumed to have only 3 vertexes. In
the future, this should be expanded to an arbitrary number of vertexes.

The model class keeps track of the minimum and maximum X, Y, and Z
coordinates, so that scaling and offsets can be easily calculated.

Vertex has a 'translate' function that applies an x y z offset and a
scaling factor.

TODO: Add unit tests for Face, Vertex, and more Model functions.

Renderer: added draw_line variant that accepts doubles instead of int.